+++ deployment/ccip/view/view.go
@@ -9,6 +9,7 @@ import (
 
 	mcmsv10 "github.com/smartcontractkit/cld-changesets/pkg/contract/mcms/view/v1_0"
 
+	"github.com/smartcontractkit/chainlink/deployment/ccip/internal/maputils"
 	"github.com/smartcontractkit/chainlink/deployment/ccip/view/aptos"
 	"github.com/smartcontractkit/chainlink/deployment/ccip/view/shared"
 	"github.com/smartcontractkit/chainlink/deployment/ccip/view/solana"
@@ -19,7 +20,6 @@ import (
 	"github.com/smartcontractkit/chainlink/deployment/ccip/view/v1_6"
 	"github.com/smartcontractkit/chainlink/deployment/common/view"
 	common_v1_0 "github.com/smartcontractkit/chainlink/deployment/common/view/v1_0"
-	"github.com/smartcontractkit/chainlink/deployment/helpers"
 )
 
 // ChainView is a json-persistable structure that represents chain state. Store all versions of CCIP contracts
@@ -119,7 +119,7 @@ func NewSolChain() SolChainView {
 func (v *ChainView) UpdateTokenPool(tokenSymbol string, tokenPoolAddress string, poolView v1_5_1.PoolView) {
 	v.UpdateMu.Lock()
 	defer v.UpdateMu.Unlock()
-	v.TokenPools = helpers.AddValueToNestedMap(v.TokenPools, tokenSymbol, tokenPoolAddress, poolView)
+	v.TokenPools = maputils.AddValueToNestedMap(v.TokenPools, tokenSymbol, tokenPoolAddress, poolView)
 }
 
 func (v *ChainView) UpdateRegistryModuleView(registryModuleAddress string, registryModuleView shared.RegistryModulesView) {
